Initial upload
A quick demo project for the adafruit Circuit Playground Express. 

If the device is mounted on a Roomba like vacuum it will generate reaction sounds in response to impacts.  One can also use the "A" and "B" buttons to adjust the sensitivity of the shake sensor.  You can install the code and test it without needing a Roomba.  

Note: If you plan on installing this on your hacked Roomba be warned. I had sorted out how to wire into the Roomba's own speaker and was in the process of testing drawing power for the board from the Roomba itself (using a 15V to 5V step-down resistor circuit) but I still managed to fry my board.